requirements.txt
# Core FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Database
asyncpg>=0.28.0

# HTTP and networking
httpx>=0.27.0
requests>=2.31.0
aiohttp>=3.8.4

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
email-validator>=2.0.0

# File handling
python-multipart>=0.0.6
jinja2>=3.1.2

# LLM API clients
openai>=1.52.0
anthropic>=0.5.0
google-generativeai>=0.3.0

# Streaming and async utilities
sse-starlette>=1.6.0
tenacity>=8.2.2

# Code processing and analysis
pygments>=2.15.0
pylint>=2.17.0
black>=23.3.0
isort>=5.12.0
mypy>=1.3.0
bandit>=1.7.5

# Token counting and utilities
tiktoken>=0.4.0

# Monitoring and logging
prometheus-client>=0.16.0
python-json-logger>=2.0.7

pytest>=7.4.0
pytest-asyncio>=0.21.0

aiofiles>=23.2.1
ujson>=5.9.0
asyncio-throttle>=1.0.2
structlog>=23.2.0
psutil>=5.9.0
cryptography>=41.0.0
gunicorn>=21.2.0
google-cloud-aiplatform>=1.38.0
redis>=4.5.0

httpx>=0.24.0
aiofiles>=0.8.0
python-multipart>=0.0.5
